+++ app/structures/role-message.js
@@ -1,8 +1,6 @@
 'use strict'
 const BaseStructure = require('./base')
-
-const { Constants } = require('discord.js')
-const { PartialTypes } = Constants
+const Postable = require('./interfaces/postable')
 
 class RoleMessage extends BaseStructure {
   constructor (client, data, guild) {
@@ -22,25 +20,12 @@ class RoleMessage extends BaseStructure {
     this.channelId = data.message.channelId
   }
 
-  get channel () {
-    return this.guild.channels.cache.get(this.channelId)
-  }
-
   get emoji () {
     return this.guild.emojis.cache.get(this._emojiId) || this._emoji
   }
 
   get emojiId () {
     return this._emoji || this._emojiId
-  }
-
-  get message () {
-    return this.messageId !== null
-      ? this.channel?.messages.cache.get(this.messageId) ||
-      (this.client.options.partials.includes(PartialTypes.MESSAGE)
-        ? this.channel?.messages.add({ id: this.messageId })
-        : null)
-      : null
   }
 
   get role () {
@@ -51,5 +36,7 @@ class RoleMessage extends BaseStructure {
     return this.guild.roleMessages.delete(this)
   }
 }
+
+Postable.applyToClass(RoleMessage)
 
 module.exports = RoleMessage
